+++ websocketpp/base64/base64.hpp
@@ -63,7 +63,7 @@ static inline bool is_base64(unsigned char c) {
  * @param len The length of input in bytes
  * @return A base64 encoded string representing input
  */
-inline std::string base64encode(unsigned char const * input, size_t len) {
+inline std::string base64_encode(unsigned char const * input, size_t len) {
     std::string ret;
     int i = 0;
     int j = 0;
@@ -116,8 +116,8 @@ inline std::string base64encode(unsigned char const * input, size_t len) {
  * @param input The input data
  * @return A base64 encoded string representing input
  */
-inline std::string base64encode(std::string const & input) {
-    return base64encode(
+inline std::string base64_encode(std::string const & input) {
+    return base64_encode(
         reinterpret_cast<const unsigned char *>(input.data()),
         input.size()
     );